+++ Assets/Mirage/Runtime/ClientObjectManager.cs
@@ -95,20 +95,20 @@ namespace Mirage
 
         internal void RegisterHostHandlers()
         {
-            Client.Player.RegisterHandler<ObjectDestroyMessage>(OnHostClientObjectDestroy);
-            Client.Player.RegisterHandler<ObjectHideMessage>(msg => { });
-            Client.Player.RegisterHandler<SpawnMessage>(OnHostClientSpawn);
-            Client.Player.RegisterHandler<ServerRpcReply>(OnServerRpcReply);
-            Client.Player.RegisterHandler<RpcMessage>(OnRpcMessage);
+            Client.MessageHandler.RegisterHandler<ObjectDestroyMessage>(OnHostClientObjectDestroy);
+            Client.MessageHandler.RegisterHandler<ObjectHideMessage>(msg => { });
+            Client.MessageHandler.RegisterHandler<SpawnMessage>(OnHostClientSpawn);
+            Client.MessageHandler.RegisterHandler<ServerRpcReply>(OnServerRpcReply);
+            Client.MessageHandler.RegisterHandler<RpcMessage>(OnRpcMessage);
         }
 
         internal void RegisterMessageHandlers()
         {
-            Client.Player.RegisterHandler<ObjectDestroyMessage>(OnObjectDestroy);
-            Client.Player.RegisterHandler<ObjectHideMessage>(OnObjectHide);
-            Client.Player.RegisterHandler<SpawnMessage>(OnSpawn);
-            Client.Player.RegisterHandler<ServerRpcReply>(OnServerRpcReply);
-            Client.Player.RegisterHandler<RpcMessage>(OnRpcMessage);
+            Client.MessageHandler.RegisterHandler<ObjectDestroyMessage>(OnObjectDestroy);
+            Client.MessageHandler.RegisterHandler<ObjectHideMessage>(OnObjectHide);
+            Client.MessageHandler.RegisterHandler<SpawnMessage>(OnSpawn);
+            Client.MessageHandler.RegisterHandler<ServerRpcReply>(OnServerRpcReply);
+            Client.MessageHandler.RegisterHandler<RpcMessage>(OnRpcMessage);
         }
 
         bool ConsiderForSpawning(NetworkIdentity identity)