+++ Assets/Mirage/Runtime/NetworkServer.cs
@@ -134,6 +134,8 @@ namespace Mirage
 
         public NetworkWorld World { get; private set; }
 
+        public IMessageHandler MessageHandler { get; internal set; }
+
         /// <summary>
         /// This shuts down the server and disconnects all clients.
         /// </summary>
@@ -207,6 +209,7 @@ namespace Mirage
         /// <returns></returns>
         public void Listen()
         {
+            MessageHandler = new MessageBroker();
             Initialize();
 
             try
@@ -311,6 +314,7 @@ namespace Mirage
             Stopped?.Invoke();
             initialized = false;
             Active = false;
+            MessageHandler = null;
         }
 
         /// <summary>
@@ -318,7 +322,7 @@ namespace Mirage
         /// </summary>
         public virtual INetworkPlayer GetNewPlayer(Connection connection)
         {
-            return new NetworkPlayer(connection);
+            return new NetworkPlayer(connection, MessageHandler);
         }
 
         /// <summary>
@@ -333,7 +337,8 @@ namespace Mirage
                 // connection cannot be null here or conn.connectionId
                 // would throw NRE
                 Players.Add(player);
-                player.RegisterHandler<NetworkPingMessage>(Time.OnServerPing);
+                // todo is registering handler multiple times a problem?
+                MessageHandler.RegisterHandler<NetworkPingMessage>(Time.OnServerPing);
             }
         }
 
@@ -425,7 +430,7 @@ namespace Mirage
             // now process messages until the connection closes
             try
             {
-                await player.ProcessMessagesAsync();
+                await MessageHandler.ProcessMessagesAsync(player);
             }
             catch (Exception ex)
             {