+++ Assets/Tests/Runtime/ClientServer/ClientServerSetup.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Linq;
 using Cysharp.Threading.Tasks;
+using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
 
@@ -132,5 +133,11 @@ namespace Mirage.Tests.ClientServer
         });
 
         #endregion
+
+        [Test]
+        public void SetupTest()
+        {
+            Assert.Pass();
+        }
     }
 }